ARTESCA-15824: Safely handle the Artesca + Veeam mode when Artesca library is not available
