+++ src/react/endpoint/hooks.ts
@@ -1,0 +1,37 @@
+import {
+  useArtescaLibrary,
+  ArtescaLibraryNotAvailable,
+} from '../next-architecture/ui/ArtescaLibraryProvider';
+
+/**
+ * Hook that safely handles Artesca Plus Veeam mode when library is not available
+ * @returns Object containing mode and status, with safe defaults when library unavailable
+ */
+export const useArtescaPlusVeeamMode = () => {
+  const artescaLibrary = useArtescaLibrary();
+
+  if (artescaLibrary instanceof ArtescaLibraryNotAvailable) {
+    // Return safe defaults when Artesca library is not available
+    return {
+      artescaPlusVeeamDefaultOrOpenMode: null,
+      artescaPlusVeeamDefaultOrOpenModeStatus: 'success',
+      ARTESCA_PLUS_VEEAM_S3_ENDPOINT_NAME: undefined,
+    };
+  }
+
+  // Use the actual hook when library is available
+  const {
+    useArtescaPlusVeeamDefaultOrOpenMode,
+    ARTESCA_PLUS_VEEAM_S3_ENDPOINT_NAME,
+  } = artescaLibrary;
+  const {
+    artescaPlusVeeamDefaultOrOpenMode,
+    artescaPlusVeeamDefaultOrOpenModeStatus,
+  } = useArtescaPlusVeeamDefaultOrOpenMode();
+
+  return {
+    artescaPlusVeeamDefaultOrOpenMode,
+    artescaPlusVeeamDefaultOrOpenModeStatus,
+    ARTESCA_PLUS_VEEAM_S3_ENDPOINT_NAME,
+  };
+};
